styling updates
